DWORD forever

+++ Sources/Overlays/_Testing_Foundation/Attachments/_AttachableURLWrapper.swift
@@ -164,7 +164,15 @@ extension _AttachableURLWrapper: FileClonable {
     // wide use at this time. SEE: https://learn.microsoft.com/en-us/windows/win32/fileio/block-cloning
     return filePath.withCString(encodedAs: UTF16.self) { filePath in
       guard let srcHandle = _fileHandle?._handle,
-            let dstHandle = CreateFileW(filePath, GENERIC_READ | GENERIC_WRITE, DWORD(FILE_SHARE_DELETE), nil, DWORD(CREATE_ALWAYS), DWORD(FILE_ATTRIBUTE_NORMAL), nil) else {
+            let dstHandle = CreateFileW(
+              filePath,
+              DWORD(GENERIC_READ | GENERIC_WRITE),
+              DWORD(FILE_SHARE_DELETE),
+              nil,
+              DWORD(CREATE_ALWAYS),
+              DWORD(FILE_ATTRIBUTE_NORMAL),
+              nil
+            ) else {
         return false
       }
       defer {
